+++ core/src/test/java/com/opensymphony/xwork2/beans/ognl/OgnlList.java
@@ -1,0 +1,44 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package com.opensymphony.xwork2.beans.ognl;
+
+import com.opensymphony.xwork2.XWorkException;
+
+import java.util.ArrayList;
+
+public class OgnlList extends ArrayList {
+    private Class clazz;
+
+    public OgnlList(Class clazz) {
+        this.clazz = clazz;
+    }
+
+    @Override
+    public synchronized Object get(int index) {
+        while (index >= this.size()) {
+            try {
+                this.add(clazz.newInstance());
+            } catch (Exception e) {
+                throw new XWorkException(e);
+            }
+        }
+
+        return super.get(index);
+    }
+}
